Split proto extensions file
